requirements.txt
albucore==0.0.13
albumentations==1.4.10
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
asyncpg==0.30.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.2
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.11
databases==0.9.0
decorator==5.1.1
dnspython==2.6.1
fastapi==0.115.6
filelock==3.16.1
fire==0.7.0
firebase-admin==6.6.0
Flask==3.0.3
Flask-Cors==4.0.1
flask-restx==1.3.0
fonttools==4.55.2
fsspec==2024.12.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.1
imgaug==0.4.0
importlib_resources==6.4.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pymongo==4.8.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.32.1
Werkzeug==3.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-cloud-texttospeech==2.24.0
groq==0.5.0
reportlab==4.0.9
duckduckgo-search==4.4.2
streamlit-authenticator
google-auth-oauthlib
google-auth
httpx-oauth
